pipeline/requirements.txt
numpy
pandas
PyYAML
pulp
h5py
tables
matplotlib
